+++ stroom-core-shared/src/main/java/stroom/pipeline/shared/XsltResource.java
@@ -31,7 +31,6 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
 
 @Api(value = "xslt - /v1")
 @Path("/xslt" + ResourcePaths.V1)
@@ -54,4 +53,14 @@ public interface XsltResource extends RestResource, DirectRestService {
             value = "Update an xslt doc",
             response = XsltDoc.class)
     XsltDoc update(XsltDoc xslt);
+
+    @GET
+    @Path("/{xsltId}")
+    XsltDoc fetch(@PathParam("xsltId") final String xsltId);
+
+    @POST
+    @Path("/{xsltId}")
+    @Consumes(MediaType.APPLICATION_JSON)
+    void save(@PathParam("xsltId") final String xsltId,
+                         final XsltDTO xsltDto);
 }