+++ stroom-pipeline/src/main/java/stroom/pipeline/xslt/XsltResourceImpl.java
@@ -19,21 +19,32 @@ package stroom.pipeline.xslt;
 import com.codahale.metrics.health.HealthCheck.Result;
 import stroom.docref.DocRef;
 import stroom.docstore.api.DocumentResourceHelper;
+import stroom.pipeline.shared.XsltDTO;
 import stroom.pipeline.shared.XsltDoc;
 import stroom.pipeline.shared.XsltResource;
+import stroom.security.api.SecurityContext;
 import stroom.util.HasHealthCheck;
 
 import javax.inject.Inject;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.core.MediaType;
 
 class XsltResourceImpl implements XsltResource, HasHealthCheck {
     private final XsltStore xsltStore;
     private final DocumentResourceHelper documentResourceHelper;
+    private SecurityContext securityContext;
 
     @Inject
     XsltResourceImpl(final XsltStore xsltStore,
-                     final DocumentResourceHelper documentResourceHelper) {
+                     final DocumentResourceHelper documentResourceHelper,
+                     final SecurityContext securityContext) {
         this.xsltStore = xsltStore;
         this.documentResourceHelper = documentResourceHelper;
+        this.securityContext = securityContext;
     }
 
     @Override
@@ -46,8 +57,48 @@ class XsltResourceImpl implements XsltResource, HasHealthCheck {
         return documentResourceHelper.update(xsltStore, doc);
     }
 
+
+
+    @GET
+    @Path("/{xsltId}")
+    public XsltDoc fetch(@PathParam("xsltId") final String xsltId) {
+        return securityContext.secureResult(() -> {
+            final XsltDoc xsltDoc = xsltStore.readDocument(getDocRef(xsltId));
+            if (null != xsltDoc) {
+                return xsltDoc;
+            } else {
+                return null;
+            }
+        });
+    }
+
+    @POST
+    @Path("/{xsltId}")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public void save(@PathParam("xsltId") final String xsltId,
+                     final XsltDTO xsltDto) {
+        // A user should be allowed to read pipelines that they are inheriting from as long as they have 'use' permission on them.
+        securityContext.useAsRead(() -> {
+            final XsltDoc xsltDoc = xsltStore.readDocument(getDocRef(xsltId));
+
+            if (xsltDoc != null) {
+                xsltDoc.setDescription(xsltDto.getDescription());
+                xsltDoc.setData(xsltDto.getData());
+                xsltStore.writeDocument(xsltDoc);
+            }
+        });
+    }
+    
     @Override
     public Result getHealth() {
         return Result.healthy();
     }
+
+    private DocRef getDocRef(final String xsltId) {
+        return new DocRef.Builder()
+                .uuid(xsltId)
+                .type(XsltDoc.DOCUMENT_TYPE)
+                .build();
+    }
+
 }